[google_sign_in/google_identity_services] Clear-up documentation of callbacks in various APIs and uses of those APIs (#8029)
These APIs used to accept allowInterop'd callbacks. When the
implementation was migrated to use dart:js_interop, they were wrapped
with toJS. However, starting in 3.5, to be consistent with dart2wasm,
double-wrapping a function is an error. The documentation should be
clear that it accepts only Dart functions and uses should be cleaned up
to not use allowInterop and type the callbacks' parameter and return
value correctly.