[in_app_purchase] Write to the transactions update queue from the main thread (#9068)
Equeue to the channel from the main thread to avoid channel safety issues.

Background: `sendTransactionUpdate` can be called from the main thread, where it is safe to send messages on the `onTransactionsUpdated` channel. This seems to happen when the purchase code path succeeds and generates a transaction. However `sendTransactionUpdate` can also be called on non-main threads serving background executors, when transaction callbacks result from restoring purchases or other transactions are reported to the transaction subscription (eg. a purchase made outside of the app, or a refund is issued -- this can be triggered from XCode's StoreKit Test transaction log). In these cases, at least sometimes, an error is logged about writing to a channel on a non-platform thread (see the issue below).

Enqueue on the main thread to avoid channel thread-safety issues. This could be optimized further, but I believe this is safe.

Resolves flutter/flutter#166493

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.