[camera_avfoundation] Implementation swift migration - part 2 (#9007)
Migrates camera implementation as part of https://github.com/flutter/flutter/issues/119109

This PR migrates `CameraPlugin` class.

While it wasn't the goal of this PR it partially resolves the [issue](https://github.com/flutter/flutter/issues/162376) regarding switching the camera during recording. The main problem there was that the capture device wasn't actually changed because the camera name passed to the `FLTCam.setDescriptionWhileRecording` wasn't used and instead a `captureDeviceFactory` was called which returned a capture device with the same name as during the initial set up. In this pr "named" capture device factory is propagated to `FLTCam` allowing passing the new capture device name during in `setDescriptionWhileRecording`.

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.