[ci] Update repository for 3.29 release (#9270)
We accidentally never did the repo update for the 3.29 release, beyond rolling the stable pin. This does the remainder of the steps from https://github.com/flutter/flutter/blob/master/docs/ecosystem/release/Updating-Packages-repo-for-a-stable-release.md, making 3.24 our min-min version, and updating the N-1 and N-2 tests to reflect the current reality.