[video_player_android] Add note about Android platform views known issue (#9312)
Just a README update to flag the [known issue](https://github.com/flutter/flutter/issues/164899) regarding the platform view mode implementation. The issue is affecting [the optional platform view mode](https://github.com/flutter/flutter/issues/86613), but the decision is to go ahead with the release including this note. cc: @stuartmorgan-g.

I'm not exactly sure about the versioning regarding this, but I think it should be visible on pub, so bumping the version is IMO appropriate.

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.