[go_router_builder] Use a mixin instead of an extension (#9275)
first step of fixing [#106790](https://github.com/flutter/flutter/issues/106790)

This PR  changes the generation to use a mixin instead of an extension. The plan is to add the routing methods directly to `GoRouteData`. An extension won't be able to override the class implementations, so I changed it to a mixin.

I'm not sure about the versioning, so feel free to suggest changes.

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.