[camera_avfoundation] fix race condition when starting image stream on iOS (#8733)
Fixes https://github.com/flutter/flutter/issues/147702 

It's not easy to reproduce this with the example application (as shown in the linked issue). We've been able to reproduce it in our own application which makes heavy usage of Platform channels and different native plugins. After digging into the code we notice that:

The Dart code that starts listening to the event channel was being executed before initializing the stream in the native side. 

https://github.com/flutter/packages/blob/aadb4841642f7a4bccac6cb99e8c834228563ef0/packages/camera/camera_avfoundation/lib/src/avfoundation_camera.dart#L271 

https://github.com/flutter/packages/blob/aadb4841642f7a4bccac6cb99e8c834228563ef0/packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/FLTThreadSafeEventChannel.m#L31

I don't have much experience with Objective-C, so not sure how to create unit tests for that if necessary but I believe this would be an adecuate solution. I've seen other methods in the plugin that use the same strategy passing the completion object. 

- [] I added new tests to check the change I am making, or this PR is [test-exempt].
- [] All existing and new tests are passing.