[camera_android_camerax] Fix incorrectly rotated captured photos (#9498)
Changes the target rotation of captured images to match that of the default display at the time of capture. This ensures that the photo's orientation is upright. This in turn fixes the rotation of the thumbnail previewing a captured photo (the thumbnail was never rotated incorrectly, just the image itself).

Fixes https://github.com/flutter/flutter/issues/169876.

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.